added websocket module

+++ back/src/routes/user.route.js
@@ -192,35 +192,47 @@ module.exports = function(app) {
      */
 
     /**
-    * Set an interest to a user                 //TODO check si les interest existent et le user existe
+    * Set an interest to a user                 //TODO check si les interest existent
     */
     app.put("/users/:user_id/interests", authMiddleware.getUserParams, (req, res) => {
 
         const user_id = parseInt(req.params.user_id);
         const interests = req.body.interest;
 
-        console.log(req.body);
-
         try {
-            //delete all interests owned by the user
-            pool.query('DELETE FROM "Interest_User" WHERE user_id = $1',
+
+
+            pool.query('SELECT * FROM "User" WHERE id = $1',
             [user_id],
-            (error) => {
+            (error, results) => {
                 if (error) throw error;
+                if (results.rowCount == 1) {
+                     //delete all interests owned by the user
+                pool.query('DELETE FROM "Interest_User" WHERE user_id = $1',
+                [user_id],
+                (error) => {
+                    if (error) throw error;
 
-                //add all the new interests
-
-                for (let index = 0; index < interests.length; index++) {
-                    const element = interests[index];
-                    pool.query('INSERT INTO "Interest_User" (user_id, interest_id) VALUES ($1, $2)',
-                    [user_id, element],
-                    (error) => {
-                        if (error) throw error;
-                    });
-                }
-                res.status(200).json({ message: 'interests updated' });
+                    //add all the new interests
+                    for (let index = 0; index < interests.length; index++) {
+                        const element = interests[index];
+                        pool.query('INSERT INTO "Interest_User" (user_id, interest_id) VALUES ($1, $2)',
+                        [user_id, element],
+                        (error) => {
+                            if (error) throw error;
+                        });
+                    }
+                    res.status(200).json({ message: 'interests updated' });
                 
+                });
+                } else {
+                    res.status(404).json({ message: 'user not found' });
+                }
             });
+
+
+
+           
         }
         catch (error) {
             console.error(error);
